updates for the clients
