+++ src/JsonRpc.Testing/JsonRpcServerTestBase.cs
@@ -36,34 +36,33 @@ namespace OmniSharp.Extensions.JsonRpc.Testing
             var clientPipe = new Pipe(TestOptions.DefaultPipeOptions);
             var serverPipe = new Pipe(TestOptions.DefaultPipeOptions);
 
-            var clientTask = JsonRpcServer.From(
-                options => {
-                    ConfigureClientInputOutput(serverPipe.Reader, clientPipe.Writer, options);
-                    clientOptionsAction(options);
-                },
-                services => services.AddTransient(typeof(IPipelineBehavior<,>), typeof(SettlePipeline<,>))
+            var clientTask = JsonRpcServer.From(options => {
+                options
+                    .WithServices(services => services
+                    .AddTransient(typeof(IPipelineBehavior<,>), typeof(SettlePipeline<,>))
                     .AddSingleton(ServerEvents as IRequestSettler)
                     .AddLogging(x => {
                         x.SetMinimumLevel(LogLevel.Trace);
                         x.Services.AddSingleton(TestOptions.ClientLoggerFactory);
-                    }),
-                null,
-                CancellationToken
-            );
+                    })
+                );
+                ConfigureClientInputOutput(serverPipe.Reader, clientPipe.Writer, options);
+                clientOptionsAction(options);
+            }, CancellationToken);
 
             var serverTask = JsonRpcServer.From(options => {
-                    ConfigureServerInputOutput(clientPipe.Reader, serverPipe.Writer, options);
-                    serverOptionsAction(options);
-                },
-                services => services.AddTransient(typeof(IPipelineBehavior<,>), typeof(SettlePipeline<,>))
+                options
+                    .WithServices(services => services
+                    .AddTransient(typeof(IPipelineBehavior<,>), typeof(SettlePipeline<,>))
                     .AddSingleton(ServerEvents as IRequestSettler)
                     .AddLogging(x => {
                         x.SetMinimumLevel(LogLevel.Trace);
-                        x.Services.AddSingleton(TestOptions.ClientLoggerFactory);
-                    }),
-                null,
-                CancellationToken
-            );
+                        x.Services.AddSingleton(TestOptions.ServerLoggerFactory);
+                    })
+                );
+                ConfigureServerInputOutput(clientPipe.Reader, serverPipe.Writer, options);
+                serverOptionsAction(options);
+            }, CancellationToken);
 
             await Task.WhenAll(clientTask, serverTask);
             _client = clientTask.Result;
